Batch configuration script enhancement
This script aims to embed the configuration from dfir-orc-config repo in a DFIR-ORC binary
When only one configuration is available, it will use the default config directory.
If another directory is available, lets say you create `MyAwesomeConf` directory next to config directory, you will have a menu to choose which one you want to build. 

As we all seem to build our own configuration, I added the option and use it all days, if it can help others ... 

I modified the script to remove the output, it will only be shown when you get an error during build process.

All Issue will be corrected if you get some.